Experimental blocking behaviour of send/subscribe/quit
Using a mutex and a 'binary semaphore' this branch implements a blocking
API. The client thread first grabs the mutex, then configures the
async_t and triggeres it and then calls sem_wait. The event loop thread,
once the async_t callback fires, does the work needed and then calls
sem_post on the semaphore, releasing the waiting client thread. Then the
client thread releases the mutex and continues.

To support this change, the 'event' callback is now run in the libuv
worker pool, rather than the event loop, otherwise a send() done from
the event callback would deadlock.