Make crypto required, don't log messages that look like errors
This mostly makes the windows CMake GUI more tolerable to use. Crypto
vars not being required and flagged as advanced made them a pain to fix.
Also the logging we were doing was coming out in red error text in the
CMake GUI, which made it look like something was wrong when it was not.